chore: move embedpaginator
